Move pedersen hash files to crypto package
