+++ core/crypto/pedersen_hash.go
@@ -1,4 +1,4 @@
-package core
+package crypto
 
 import (
 	"errors"
