+++ core/crypto/pedersen_hash_test.go
@@ -1,4 +1,4 @@
-package core
+package crypto
 
 import (
 	"fmt"
